Could generate tortuosity on-the-fly with this commit.
Updated:
- solver settings for Study 1 so only the "correct" variables are run
- added code to add the Dilute Species Transport and its B.C.s, I.C.s
- added code to add the study for the tortuosity study
- added function to calculate flux to compute tortuosity

Note:
Somehow broke the boundary probe plot. When I checked after running the
simulation, there weren't any voltage values in the plot. But if I plot
solid or electrolyte concentration, stuff pops up, so there are values.

However, by running the study manually, I get the voltage values in the table
and in the plot.

Could be something to look into if there's a need for these probe values.